fix header menu and calendar UI bugs
