+++ src/components/Header/FilterMenu.tsx
@@ -3,7 +3,8 @@ import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 import Button from "@material-ui/core/Button";
 import Menu from "@material-ui/core/Menu";
 import MenuItem from "@material-ui/core/MenuItem";
-import { makeStyles } from "@material-ui/core";
+import { makeStyles, withStyles } from "@material-ui/core";
+import styled from "styled-components";
 
 const useStyle = makeStyles((theme) => ({
     button: {
@@ -15,8 +16,20 @@ const useStyle = makeStyles((theme) => ({
         float: "right",
         marginLeft: "20px",
         fontSize: "16px",
+        fontFamily: "Source Sans Pro",
+    },
+    menuItem: {
+        fontFamily: "Source Sans Pro",
+        padding: "6px 50px",
     },
 }));
+
+const StyledMenu = styled(({ className, ...props }) => (
+    <Menu classes={{ paper: className }} {...props} />
+))`
+    box-shadow: 0px 4px 20px rgba(80, 80, 80, 0.25);
+    border-radius: 12px;
+`;
 
 const FilterMenu = ({ options, selected, setSelected }: any) => {
     const classes = useStyle();
@@ -47,15 +60,26 @@ const FilterMenu = ({ options, selected, setSelected }: any) => {
                 {options[selected]}
                 <ExpandMoreIcon />
             </Button>
-            <Menu
+            <StyledMenu
                 id="filter-menu"
                 anchorEl={anchorEl}
+                elevation={0}
+                getContentAnchorEl={null}
+                anchorOrigin={{
+                    vertical: "bottom",
+                    horizontal: "center",
+                }}
+                transformOrigin={{
+                    vertical: "top",
+                    horizontal: "center",
+                }}
                 keepMounted
                 open={Boolean(anchorEl)}
                 onClose={handleClose}
             >
                 {Object.keys(options).map((key: string) => (
                     <MenuItem
+                        className={classes.menuItem}
                         key={key}
                         selected={key === selected}
                         onClick={(e) => handleSelect(e, key)}
@@ -63,7 +87,7 @@ const FilterMenu = ({ options, selected, setSelected }: any) => {
                         {options[key]}
                     </MenuItem>
                 ))}
-            </Menu>
+            </StyledMenu>
         </div>
     );
 };
