+++ src/components/TrashButton/index.tsx
@@ -1,25 +1,5 @@
 import React from "react";
-import styled from "styled-components";
-
-const DeleteButton = styled.div`
-    background: #5094b9;
-    border-radius: 11px;
-    border: none;
-    padding: 4px;
-    width: 30px;
-    height: 30px;
-    display: flex;
-    font-size: 18px;
-    justify-content: center;
-    align-items: center;
-    :hover {
-        cursor: pointer;
-    }
-    margin: 0px;
-    * {
-        margin: 0px;
-    }
-`;
+import { DeleteButton } from "./style";
 
 const TrashButton = ({ onClick }: any) => {
     return (
