+++ src/components/TrashButton/style.ts
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const DeleteButton = styled.div`
+    background: #5094b9;
+    border-radius: 11px;
+    border: none;
+    padding: 4px;
+    width: 30px;
+    height: 30px;
+    display: flex;
+    font-size: 18px;
+    justify-content: center;
+    align-items: center;
+    :hover {
+        cursor: pointer;
+    }
+    margin: 0px;
+    * {
+        margin: 0px;
+    }
+`;
