+++ src/components/WidgetPaper/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { StyledBackground } from "./style";
+
+const WidgetPaper = ({ children }: any) => {
+    return <StyledBackground>{children}</StyledBackground>;
+};
+
+export default WidgetPaper;
